requirements.txt
pywebview~=5.1
pyinstaller
python-dotenv~=1.0.1